refactor(service): status service logic to profile service
