--- server/src/main/java/pl/talkapp/server/service/status/StatusService.java
@@ -1,11 +1,0 @@
-package pl.talkapp.server.service.status;
-
-import pl.talkapp.server.entity.Status;
-
-public interface StatusService {
-    Status getOnline();
-
-    Status getOffline();
-
-    Status getBusy();
-}
