+++ server/src/main/java/pl/talkapp/server/service/user/UserProfileServiceImpl.java
@@ -2,8 +2,13 @@ package pl.talkapp.server.service.user;
 
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import pl.talkapp.server.entity.Status;
 import pl.talkapp.server.entity.User;
+import pl.talkapp.server.repository.StatusRepository;
 import pl.talkapp.server.repository.UserRepository;
+
+import java.util.Map;
+import java.util.stream.Collectors;
 
 @Service
 public class UserProfileServiceImpl implements UserProfileService {
@@ -11,9 +16,14 @@ public class UserProfileServiceImpl implements UserProfileService {
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
 
-    public UserProfileServiceImpl(UserRepository userRepository, PasswordEncoder passwordEncoder) {
+    Map<pl.talkapp.server.model.Status, Status> statuses;
+
+    public UserProfileServiceImpl(UserRepository userRepository, PasswordEncoder passwordEncoder,
+                                  StatusRepository statusRepository) {
         this.userRepository = userRepository;
         this.passwordEncoder = passwordEncoder;
+        statuses = statusRepository.findAll().stream().collect(Collectors.toMap(Status::getName,
+                status -> status));
     }
 
     @Override
@@ -30,6 +40,27 @@ public class UserProfileServiceImpl implements UserProfileService {
     @Override
     public void changeName(User user, String nick) {
         user.setName(nick);
+        userRepository.save(user);
+    }
+
+    @Override
+    public Status getOnline() {
+        return statuses.get(pl.talkapp.server.model.Status.ONLINE);
+    }
+
+    @Override
+    public Status getOffline() {
+        return statuses.get(pl.talkapp.server.model.Status.OFFLINE);
+    }
+
+    @Override
+    public Status getBusy() {
+        return statuses.get(pl.talkapp.server.model.Status.BUSY);
+    }
+
+    @Override
+    public void setStatus(User user, Status status) {
+        user.setStatus(status);
         userRepository.save(user);
     }
 }
